Use puppet/nodejs module instead of willdurand/nodejs
The verdaccio module currently relies on version 1.9.x of willdurand/nodejs,
although this is not explicitly specified. In the meantime this branch has
become deprecated and, although there is a 2.0 version, the puppet/nodejs
module has become the 'de facto' standard.

By changing the manifest to use the puppet/nodejs module (and explicitly
stating it as a dependency), it becomes easier for others to use the
verdaccio module and not having it clash with the nodejs module they
are probably already using.